+++ src/aux-auth/serverless/aws/src/ServerlessRecordsManager.ts
@@ -37,6 +37,11 @@ export interface GetRecordsResult {
     data?: GetRecordsActionResult;
 }
 
+export interface DeleteRecordResult {
+    status: number;
+    message?: string;
+}
+
 export interface ServerlessGetRecordsRequest {
     token?: string;
     issuer: string;
@@ -45,6 +50,15 @@ export interface ServerlessGetRecordsRequest {
     cursor?: string;
     space: RecordSpace;
 }
+
+export interface ServerlessDeleteRecordRequest {
+    token: string;
+    issuer: string;
+    address: string;
+    space: RecordSpace;
+}
+
+const _24_HOURS_IN_MILISECONDS = 1000 * 60 * 60 * 24;
 
 export class ServerlessRecordsManager {
     private _auth: AuthProvider;
@@ -133,35 +147,21 @@ export class ServerlessRecordsManager {
     async getRecords(
         request: ServerlessGetRecordsRequest
     ): Promise<GetRecordsResult> {
+        let tokenResult = this._parseToken(request.token);
+
         let tokenIssuer: string;
         let bundle: string;
-
-        if (hasValue(request.token)) {
-            const tokenResult = parseAuthToken(request.token);
-
-            if (!tokenResult) {
-                return {
-                    status: 403,
-                    message: 'Invalid token.',
-                };
-            }
-
-            const [token, b] = tokenResult;
-
-            const issuer = this._auth.validateToken(token, b);
-
-            if (!issuer) {
-                return {
-                    status: 403,
-                    message: 'Invalid token.',
-                };
-            }
-
-            tokenIssuer = issuer;
-            bundle = b;
+        if (tokenResult === null) {
+            return {
+                status: 403,
+                message: 'Invalid token.',
+            };
+        } else if (!!tokenResult) {
+            tokenIssuer = tokenResult[0];
+            bundle = tokenResult[1];
         }
 
-        let authToken =
+        const authToken =
             hasValue(tokenIssuer) && hasValue(bundle)
                 ? formatAuthToken(tokenIssuer, bundle)
                 : undefined;
@@ -239,5 +239,106 @@ export class ServerlessRecordsManager {
             status: 500,
             message: 'A server error occurred.',
         };
+    }
+
+    async deleteRecord(
+        request: ServerlessDeleteRecordRequest
+    ): Promise<DeleteRecordResult> {
+        if (!hasValue(request.token)) {
+            return {
+                status: 400,
+                message: 'Invalid request. A auth token must be provided.',
+            };
+        }
+
+        const tokenResult = this._parseToken(request.token);
+
+        if (!tokenResult) {
+            return {
+                status: 403,
+                message: 'Invalid token.',
+            };
+        }
+
+        const [issuer, bundle, expireTime] = tokenResult;
+
+        if (expireTime - _24_HOURS_IN_MILISECONDS > Date.now()) {
+            return {
+                status: 403,
+                message: 'Permanent auth tokens cannot delete records.',
+            };
+        }
+
+        if (
+            !hasValue(issuer) ||
+            !hasValue(bundle) ||
+            !hasValue(request.space) ||
+            !hasValue(request.address)
+        ) {
+            return {
+                status: 400,
+                message: 'Invalid request.',
+            };
+        }
+
+        let deleted: boolean;
+        if (
+            request.space === 'permanentGlobal' ||
+            request.space === 'permanentRestricted'
+        ) {
+            await this._store.deletePermanentRecord({
+                issuer,
+                address: request.address,
+            });
+            deleted = true;
+        } else if (
+            request.space === 'tempGlobal' ||
+            request.space === 'tempRestricted'
+        ) {
+            await this._store.deleteTemporaryRecord({
+                issuer,
+                address: request.address,
+            });
+            deleted = true;
+        }
+
+        if (deleted) {
+            return {
+                status: 200,
+            };
+        }
+
+        console.error(
+            '[ServerlessRecordsManager] Did not handle request for space: ' +
+                request.space
+        );
+        return {
+            status: 500,
+            message: 'A server error occurred.',
+        };
+    }
+
+    private _parseToken(requestToken: string) {
+        if (hasValue(requestToken)) {
+            const tokenResult = parseAuthToken(requestToken);
+
+            if (!tokenResult) {
+                return null;
+            }
+
+            const [token, b] = tokenResult;
+
+            const issuer = this._auth.validateToken(token, b);
+
+            if (!issuer) {
+                return null;
+            }
+
+            const expireTime = this._auth.getTokenExpireTime(token);
+
+            return [issuer, b, expireTime] as const;
+        }
+
+        return undefined;
     }
 }