+++ src/aux-auth/serverless/aws/src/handlers/records.js
@@ -69,21 +69,37 @@ function returnResult(event, result) {
     );
 }
 
-async function postRecord(event) {
+async function postOrDeleteRecord(event) {
     if (event.httpMethod !== 'POST') {
         throw new Error(
-            `postRecord only accept POST method, you tried: ${event.httpMethod}`
+            `postOrDeleteRecord only accept POST method, you tried: ${event.httpMethod}`
         );
     }
 
+    if (event.path.endsWith('/delete')) {
+        return deleteRecord(event);
+    } else {
+        return postRecord(event);
+    }
+}
+
+async function deleteRecord(event) {
     if (!validateOrigin(event, allowedOrigins)) {
         throw new Error('Invalid origin');
     }
 
     const data = JSON.parse(event.body);
+    const result = await manager.deleteRecord(data);
+    return returnResult(event, result);
+}
 
+async function postRecord(event) {
+    if (!validateOrigin(event, allowedOrigins)) {
+        throw new Error('Invalid origin');
+    }
+
+    const data = JSON.parse(event.body);
     const result = await manager.publishRecord(data);
-
     return returnResult(event, result);
 }
 
@@ -143,7 +159,7 @@ async function getRecords(event) {
 
 export async function handleRecords(event) {
     if (event.httpMethod === 'POST') {
-        return await postRecord(event);
+        return await postOrDeleteRecord(event);
     } else {
         return await getRecords(event);
     }