+++ src/aux-common/runtime/AuxLibrary.ts
@@ -247,6 +247,10 @@ import {
     GetRecordsAction,
     GetRecordsActionResult,
     GetRecordsQuery,
+    requestPermanentAuthToken as calcRequestPermanentAuthToken,
+    PermanentAuthTokenResult,
+    DeletableRecord,
+    deleteRecord,
 } from '../bots';
 import { sortBy, every } from 'lodash';
 import {
@@ -852,9 +856,11 @@ export function createDefaultLibrary(context: AuxGlobalContext) {
                 unregisterApp,
                 compileApp: setAppContent,
                 requestAuthBot,
+                requestPermanentAuthToken,
 
                 publishRecord,
                 getRecords,
+                destroyRecord,
             },
 
             portal: {
@@ -2685,14 +2691,25 @@ export function createDefaultLibrary(context: AuxGlobalContext) {
     }
 
     /**
+     * Requests an auth token that does not expire and can be used to authorize other app bundles to publish records for this app bundle.
+     */
+    async function requestPermanentAuthToken(): Promise<string> {
+        const task = context.createTask();
+        const event = calcRequestPermanentAuthToken(task.taskId);
+        const data: PermanentAuthTokenResult = await addAsyncAction(
+            task,
+            event
+        );
+        return formatAuthToken(data.token, data.service);
+    }
+
+    /**
      * Publishes a record that can be used across servers.
      * @param recordDefinition The data that should be used to publish the record.
      */
     function publishRecord(
         recordDefinition: PublishableRecord
     ): Promise<RecordReference> {
-        const task = context.createTask();
-
         let address: string;
         if ('address' in recordDefinition) {
             if (!hasValue(recordDefinition.address)) {
@@ -2728,6 +2745,7 @@ export function createDefaultLibrary(context: AuxGlobalContext) {
             throw new Error('The record property is required.');
         }
 
+        const task = context.createTask();
         const event = calcPublishRecord(
             token,
             address,
@@ -2824,6 +2842,38 @@ export function createDefaultLibrary(context: AuxGlobalContext) {
                 },
             };
         }
+    }
+
+    /**
+     * Requests that the given record be destroyed.
+     * @param record The record that should be deleted.
+     */
+    function destroyRecord(record: DeletableRecord) {
+        let address: string;
+        if (!hasValue(record.address)) {
+            throw new Error(
+                'the address property is required in order to delete a record.'
+            );
+        }
+        address = record.address;
+
+        if (!hasValue(record.space)) {
+            throw new Error(
+                'the space property is required in order to delete a record.'
+            );
+        }
+        const space = record.space;
+
+        const token =
+            record.authToken ?? (<any>globalThis).authBot?.tags?.authToken;
+
+        if (!hasValue(token)) {
+            throw new Error('authToken is required when there is no authBot.');
+        }
+
+        const task = context.createTask();
+        const event = deleteRecord(token, address, space, task.taskId);
+        return addAsyncAction(task, event);
     }
 
     /**