+++ src/aux-server/aux-web/aux-player/PlayerApp/PlayerApp.ts
@@ -978,6 +978,18 @@ export default class PlayerApp extends Vue {
                             asyncError(e.taskId, ex.toString())
                         );
                     }
+                } else if (e.type === 'request_permanent_auth_token') {
+                    try {
+                        const data = await simulation.auth.getPermanentAuthToken();
+
+                        simulation.helper.transaction(
+                            asyncResult(e.taskId, data, false)
+                        );
+                    } catch (ex) {
+                        simulation.helper.transaction(
+                            asyncError(e.taskId, ex.toString())
+                        );
+                    }
                 }
             }),
             simulation.connection.connectionStateChanged.subscribe(